Encrypt or decrypt all files in a directory

WARNING: This function must be used with caution! Check the parameters carefully.

lock_dir() and unlock_dir() can encrypt or decrypt all files from a given directory.

lock_dir(
  dir = "./inst/extdata",
  public_key = "./inst/ssh/id_rsa.pub",
  suffix = ".lockr",
  remove_file = TRUE
)

unlock_dir(
  dir = "./inst/extdata",
  private_key = "./inst/ssh/id_rsa",
  suffix = ".lockr",
  remove_file = TRUE,
  password = NULL
)

Arguments

dir

A string indicating the directory to encrypt/decrypt. (default: "./inst/extdata").

public_key

(optional) an openssl RSA public key or a string specifying the public key path. See rsa_keygen() to learn how to create an RSA key pair (default: "./inst/ssh/id_rsa.pub").

suffix

(optional) a string specifying the suffix to be added (when encrypting)/removed (when decrypting) to/of the file. It must start with . (default: ".lockr").

remove_file

(optional) a logical value indicating if the original file must be removed after the encryption/decryption process (default: TRUE).

private_key

(optional) an openssl RSA private key or a string specifying the private key path. See rsa_keygen() to learn how to create an RSA key pair (default: "./inst/ssh/id_rsa").

password

(optional) only for protected keys. A string specifying the password to read the private key. Avoid typing passwords on the console, use askpass() instead (default: NULL).

Value

An invisible NULL. These functions are called just for side effects.

See also

Examples

## Locking files

ssh_dir <- tempfile("ssh")
dir.create(ssh_dir)
rsa_keygen(ssh_dir)
#> Keys successfully created at '/tmp/RtmpCbw9TB/ssh164d4a1c27a5'.

temp_dir <- tempfile("dir")
dir.create(temp_dir)
for (i in seq(5)) {
    file.create(tempfile(tmpdir = temp_dir))
}
list.files(temp_dir)
#> [1] "file164d1486e652" "file164d1d712e5d" "file164d27f36dd1" "file164d450015b" 
#> [5] "file164dc74b865" 

lock_dir(temp_dir, public_key = file.path(ssh_dir, "id_rsa.pub"))
#> Locked file written at
#> '/tmp/RtmpCbw9TB/dir164d590e5053/file164d1486e652.lockr'.
#> Locked file written at
#> '/tmp/RtmpCbw9TB/dir164d590e5053/file164d1d712e5d.lockr'.
#> Locked file written at
#> '/tmp/RtmpCbw9TB/dir164d590e5053/file164d27f36dd1.lockr'.
#> Locked file written at '/tmp/RtmpCbw9TB/dir164d590e5053/file164d450015b.lockr'.
#> Locked file written at '/tmp/RtmpCbw9TB/dir164d590e5053/file164dc74b865.lockr'.

## Unlocking files

unlock_dir(temp_dir, private_key = file.path(ssh_dir, "id_rsa"))
#> Unlocked file written at '/tmp/RtmpCbw9TB/dir164d590e5053/file164d1486e652'.
#> Unlocked file written at '/tmp/RtmpCbw9TB/dir164d590e5053/file164d1d712e5d'.
#> Unlocked file written at '/tmp/RtmpCbw9TB/dir164d590e5053/file164d27f36dd1'.
#> Unlocked file written at '/tmp/RtmpCbw9TB/dir164d590e5053/file164d450015b'.
#> Unlocked file written at '/tmp/RtmpCbw9TB/dir164d590e5053/file164dc74b865'.