Encrypt or decrypt single files

lock_file() and unlock_file can encrypt/decrypt any kind of file using an RSA key pair. If your project doesn't have an RSA key pair, you can use rsa_keygen() to create one.

These functions use encrypt_envelope() / decrypt_envelope() to perform file encryption/decryption. See those functions to learn more about the encrypting/decrypting process.

lock_file(
  file,
  public_key = "./inst/ssh/id_rsa.pub",
  suffix = ".lockr",
  remove_file = TRUE
)

unlock_file(
  file,
  private_key = "./inst/ssh/id_rsa",
  suffix = ".lockr",
  remove_file = TRUE,
  password = NULL
)

Arguments

file

A string with the file path to be encrypted/decrypted. For security reasons, encrypted files must end with the suffix parameter.

public_key

(optional) an openssl RSA public key or a string specifying the public key path. See rsa_keygen() to learn how to create an RSA key pair (default: "./inst/ssh/id_rsa.pub").

suffix

(optional) a string specifying the suffix to be added (when encrypting)/removed (when decrypting) to/of the file. It must start with . (default: ".lockr").

remove_file

(optional) a logical value indicating if the original file must be removed after the encryption/decryption process (default: TRUE).

private_key

(optional) an openssl RSA private key or a string specifying the private key path. See rsa_keygen() to learn how to create an RSA key pair (default: "./inst/ssh/id_rsa").

password

(optional) only for protected keys. A string specifying the password to read the private key. Avoid typing passwords on the console, use askpass() instead (default: NULL).

Value

An invisible string containing the locked/unlocked file path.

See also

Examples

## Locking files

temp_dir <- tempfile("dir")
dir.create(temp_dir)
temp_file <- tempfile(tmpdir = temp_dir)
rsa_keygen(temp_dir)
#> Keys successfully created at '/tmp/RtmpPXVuqU/dir156b276c6732'.

con <- file(temp_file, "w+")
cat("Test", file = temp_file, sep = "\n")
list.files(temp_dir)
#> [1] "file156b4ed54410" "id_rsa"           "id_rsa.pub"      
suppressWarnings(readLines(con))
#> [1] "Test"
close(con)

lock_file(temp_file, public_key = file.path(temp_dir, "id_rsa.pub"))
#> Locked file written at
#> '/tmp/RtmpPXVuqU/dir156b276c6732/file156b4ed54410.lockr'.

temp_file_locked <- paste0(temp_file, ".lockr")
con <- file(temp_file_locked, "rb")
list.files(temp_dir)
#> [1] "file156b4ed54410.lockr" "id_rsa"                 "id_rsa.pub"            
suppressWarnings(readLines(con))
#> [1] "\037\x8b\b"                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                  
#> [2] "\xae?6K\x85\xe3\xfb\x83\xe2\x9e\xf5<\"\001\x8e\xd5\xdf;.\xe5z\xb0d\037\xf7\xf5\xf4|\u1b4d\xcf3\xcd\xf6]M\xf8(~~\003\x83\x91^y\xe0\x9e'\177\031\f\xe6/\xfc\x90\xc9m\xc3%\022\xc1\xbb~\xd6\xed\037\xcb\xfe44\xfa\xda\xc9\xff\xbd\xbb\xc2$K\xf0\xfb\xcf\xe3&\xf3\xebއ\t%MiT\xfctu?\xaf\xee\xbd\006\xdd\xe3܂/˿\xcb\xf3\xda\xfd\xb6\xb9d?1\xfbK\xfe\036\x8f\003\f\x91r\xcc\021\xf2\xab;\x96f\xfc\027\xfbyb\x81\x89\x9b\xcd\xc3\a\x93\023d̪\x9f,\x92\xf2\xfe'\xa2\xae\xa7\xb1\xe6ߏK\016\xf2b\t6;y\036\xb0\xb2<\xde\xf9\xfa\x8a{\xe7\xf4\xe7>O\x96*\xfc\xbc\xf6\x98\177m\xe7u\xc6\031S\xf5D\xf7}\x93gk`i4\xbar\xd6}Q\x98l\xe37\xab\xf8\005\xf5o\002N\xca~x\xef6os\xcb)X\xf8<\x93\xd7/\xe2\\~\x9b}a\x8dn\xdd\xfb\xff&\x91\f\xc0\020e"
#> [3] "!\xa1"                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                       
#> [4] "\026d\xca,\x83\xb2؋S\x8b\x8b3\xf3\xf3\xa0\\\x96\x94ĒD \xfd\017"                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                              
close(con)

## Unlocking files

unlock_file(temp_file_locked, private_key = file.path(temp_dir, "id_rsa"))
#> Unlocked file written at '/tmp/RtmpPXVuqU/dir156b276c6732/file156b4ed54410'.

list.files(temp_dir)
#> [1] "file156b4ed54410" "id_rsa"           "id_rsa.pub"      
con <- file(temp_file, "r+")
readLines(con)
#> [1] "Test"
close(con)